+++ listings/ch09/listing_22.js
@@ -1,56 +1,23 @@
-function patchChildren(oldVdom, newVdom/*--add--*/, hostComponent/*--add--*/) {
-  const oldChildren = extractChildren(oldVdom)
-  const newChildren = extractChildren(newVdom)
-  const parentEl = oldVdom.el
+export function defineComponent({ render, state, ...methods }) {
+  const Component = class {
+    // --snip-- //
 
-  const diffSeq = arraysDiffSequence(
-    oldChildren,
-    newChildren,
-    areNodesEqual
-  )
-  
-  for (const operation of diffSeq) {
-    const { from, index, item } = operation
-
-    switch (operation.op) {
-      case ARRAY_DIFF_OP.ADD: {
-        mountDOM(item, parentEl, index/*--add--*/, hostComponent/*--add--*/)
-        break
+    mount(hostEl, index = null) {
+      if (this.#isMounted) {
+        throw new Error('Component is already mounted')
       }
-
-      case ARRAY_DIFF_OP.REMOVE: {
-        destroyDOM(item)
-        break
-      }
-        
-      case ARRAY_DIFF_OP.MOVE: {
-        const el = oldChildren[from].el
-        const elAtTargetIndex = parentEl.childNodes[index]
-
-        parentEl.insertBefore(el, elAtTargetIndex)
-        patchDOM(
-          oldChildren[from],
-          newChildren[index],
-          parentEl,
-          // --add--
-          hostComponent
-          // --add--
-        )
-
-        break
-      }
-
-      case ARRAY_DIFF_OP.NOOP: {
-        patchDOM(
-          oldChildren[from],
-          newChildren[index],
-          parentEl,
-          // --add--
-          hostComponent
-          // --add--
-        )
-        break
-      }
+      
+      this.#vdom = this.render()
+      mountDOM(this.#vdom, hostEl, index/*--add--*/, this/*--add--*/)
+      
+      this.#hostEl = hostEl
+      this.#isMounted = true
     }
+    
+    // --snip-- //
   }
+
+  // --snip-- //
+
+  return Component
 }
